+++ okta/app.go
@@ -2,12 +2,15 @@ package okta
 
 import (
 	"context"
+	"crypto/sha256"
+	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"io"
+	"log"
 	"net/http"
+	"os"
 	"sync"
-
-	"github.com/okta/terraform-provider-okta/sdk"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/okta/okta-sdk-golang/v2/okta"
@@ -87,8 +90,7 @@ var baseAppSchema = map[string]*schema.Schema{
 			if logoPath == "" {
 				return logoPath
 			}
-
-			return fmt.Sprintf("%s (%s)", logoPath, sdk.GetAppLogoHash(logoPath))
+			return fmt.Sprintf("%s (%s)", logoPath, computeFileHash(logoPath))
 		},
 	},
 	"logo_url": {
@@ -608,4 +610,19 @@ func listAppUsersAndGroupsIDs(ctx context.Context, client *okta.Client, id strin
 		groups[i] = appGroups[i].Id
 	}
 	return
+}
+
+func computeFileHash(filename string) string {
+	file, err := os.Open(filename)
+	if err != nil {
+		return ""
+	}
+	defer func() {
+		_ = file.Close()
+	}()
+	h := sha256.New()
+	if _, err := io.Copy(h, file); err != nil {
+		log.Fatal(err)
+	}
+	return hex.EncodeToString(h.Sum(nil))
 }
